backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
redis==5.0.1
pyseto==1.7.5
cryptography==41.0.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
httpx==0.28.1
aiokafka==0.12.0
requests==2.31.0
prometheus-client==0.19.0
pytest==8.0.0